fix(ui): add `@param` use-site target for annotations fix all warning
The `@DrawableRes` and `@StringRes` annotations were being applied to properties, but they should be applied to the constructor parameters.

This commit updates the annotation usage to `@param:DrawableRes` and `@param:StringRes` to correctly target the constructor parameters.

Additionally, in `DropOffScreen.kt`, the `menuAnchor` modifier was updated to explicitly set the `type` and `enabled` parameters for clarity and potentially to address any underlying issues with the default behavior.